Extracted ValidatorSetDelta tests into separate file
